problem-solver/py/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
browser-cookie3==0.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
curl-cffi==0.5.9
distro==1.9.0
duckduckgo-search==3.9.6
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==3.0.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
g4f==0.1.8.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jeepney==0.8.0
Jinja2==3.1.3
Js2Py==0.74
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
nest-asyncio==1.5.8
numpy==1.26.4
openai==0.28.0
packaging==23.2
pyttsx3==2.9.0
Pillow==10.1.0
platformdirs==4.0.0
Py-Arkose-Generator==0.0.0.2
Py-Sc-Client==0.3.1
Py-Sc-Kpm==0.2.0
PyAudio==0.2.14
PyCryptodome==3.19.0
PyCryptodomeX==3.19.0
Pydantic==2.6.3
Pydantic-Core==2.16.3
PyExecJS==1.5.1
Pyjsparser==2.7.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
SpeechRecognition==3.10.1
sympy==1.12
tokenizers==0.15.2
tqdm==4.66.2
typing-extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.24.0.post1
waitress==2.1.2
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4

